+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./client.js";
 export * from "./manifest.js";
 export * from "./reference-policy.js";
+export * from "./reference-reward.js";
 export * from "./types.js";